add: spotlesscheck and refactoring
