+++ posttest/Dockerfile
@@ -1,12 +1,13 @@
-# stage-1 run tests
-
-
-# stage-2 build artifact
+# stage-1 build artifact
 FROM amazoncorretto:17.0.9-alpine3.18 as builder
 WORKDIR /app
 ADD . .
 RUN chmod +x ./gradlew
 RUN ["./gradlew", "clean", "bootJar"]
+
+# stage-2 running unit tests
+FROM builder as test
+RUN ["./gradlew", "test"]
 
 # stage-3 running image
 FROM gcr.io/distroless/java17-debian12:latest
