+++ posttest/build.gradle
@@ -2,6 +2,14 @@ plugins {
 	id 'java'
 	id 'org.springframework.boot' version '3.2.2'
 	id 'io.spring.dependency-management' version '1.1.4'
+	id 'com.diffplug.spotless' version '6.22.0'
+}
+
+spotless {
+	java {
+		removeUnusedImports()
+		indentWithTabs()
+	}
 }
 
 group = 'com.kbtg.bootcamp'
