+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/admin/service/AdminService.java
@@ -11,33 +11,33 @@ import org.springframework.stereotype.Service;
 @Service
 public class AdminService {
 
-    private final LotteryRepository lotteryRepository;
+	private final LotteryRepository lotteryRepository;
 
-    public AdminService(LotteryRepository lotteryRepository) {
-        this.lotteryRepository = lotteryRepository;
-    }
+	public AdminService(LotteryRepository lotteryRepository) {
+		this.lotteryRepository = lotteryRepository;
+	}
 
-    @Transactional
-    public createLotteryResponse createLottery(createLotteryRequestDto request) {
+	@Transactional
+	public createLotteryResponse createLottery(createLotteryRequestDto request) {
 
-        String requestTicket = request.ticket();
+		String requestTicket = request.ticket();
 
-        // Check if the lottery number already exists
-        if (lotteryRepository.existsByTicketid(requestTicket)) {
+		// Check if the lottery number already exists
+		if (lotteryRepository.existsByTicketid(requestTicket)) {
 
-            Lottery lottery = lotteryRepository.findByTicketid(requestTicket);
-            lottery.setAmount(lottery.getAmount() + request.amount());
-            lotteryRepository.save(lottery);
+			Lottery lottery = lotteryRepository.findByTicketid(requestTicket);
+			lottery.setAmount(lottery.getAmount() + request.amount());
+			lotteryRepository.save(lottery);
 
-        }
+		}
 
-        // If the lottery number does not exist, create a new lottery
-        else {
-            Lottery newLottery = new Lottery(requestTicket, request.price(), request.amount());
-            lotteryRepository.save(newLottery);
+		// If the lottery number does not exist, create a new lottery
+		else {
+			Lottery newLottery = new Lottery(requestTicket, request.price(), request.amount());
+			lotteryRepository.save(newLottery);
 
-        }
-        return new createLotteryResponse(requestTicket);
+		}
+		return new createLotteryResponse(requestTicket);
 
-    }
+	}
 }
