+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/exception/ApiExceptionHandler.java
@@ -13,34 +13,34 @@ import java.time.ZonedDateTime;
 @ControllerAdvice
 public class ApiExceptionHandler extends ResponseEntityExceptionHandler {
 
-    @ExceptionHandler(value = {BadRequestException.class})
-    public ResponseEntity<Object> handleBadRequestException(BadRequestException badRequestException) {
-        ApiExceptionResponse apiExceptionResponse = new ApiExceptionResponse(
-                badRequestException.getMessage(),
-                HttpStatus.BAD_REQUEST,
-                ZonedDateTime.now()
-        );
-        return new ResponseEntity<>(apiExceptionResponse, HttpStatus.BAD_REQUEST);
-    }
+	@ExceptionHandler(value = {BadRequestException.class})
+	public ResponseEntity<Object> handleBadRequestException(BadRequestException badRequestException) {
+		ApiExceptionResponse apiExceptionResponse = new ApiExceptionResponse(
+				badRequestException.getMessage(),
+				HttpStatus.BAD_REQUEST,
+				ZonedDateTime.now()
+		);
+		return new ResponseEntity<>(apiExceptionResponse, HttpStatus.BAD_REQUEST);
+	}
 
-    @ExceptionHandler(value = {InternalServiceException.class})
-    public ResponseEntity<Object> handleInternalServiceException(InternalServiceException internalServiceException) {
-        ApiExceptionResponse apiExceptionResponse = new ApiExceptionResponse(
-                internalServiceException.getMessage(),
-                HttpStatus.INTERNAL_SERVER_ERROR,
-                ZonedDateTime.now()
-        );
-        return new ResponseEntity<>(apiExceptionResponse, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
+	@ExceptionHandler(value = {InternalServiceException.class})
+	public ResponseEntity<Object> handleInternalServiceException(InternalServiceException internalServiceException) {
+		ApiExceptionResponse apiExceptionResponse = new ApiExceptionResponse(
+				internalServiceException.getMessage(),
+				HttpStatus.INTERNAL_SERVER_ERROR,
+				ZonedDateTime.now()
+		);
+		return new ResponseEntity<>(apiExceptionResponse, HttpStatus.INTERNAL_SERVER_ERROR);
+	}
 
-    @ExceptionHandler(value = {NotFoundException.class})
-    public ResponseEntity<Object> handleNotFoundException(NotFoundException notFoundException) {
-        ApiExceptionResponse apiExceptionResponse = new ApiExceptionResponse(
-                notFoundException.getMessage(),
-                HttpStatus.NOT_FOUND,
-                ZonedDateTime.now()
-        );
-        return new ResponseEntity<>(apiExceptionResponse, HttpStatus.NOT_FOUND);
-    }
+	@ExceptionHandler(value = {NotFoundException.class})
+	public ResponseEntity<Object> handleNotFoundException(NotFoundException notFoundException) {
+		ApiExceptionResponse apiExceptionResponse = new ApiExceptionResponse(
+				notFoundException.getMessage(),
+				HttpStatus.NOT_FOUND,
+				ZonedDateTime.now()
+		);
+		return new ResponseEntity<>(apiExceptionResponse, HttpStatus.NOT_FOUND);
+	}
 
 }
