+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/exception/ApiExceptionResponse.java
@@ -1,30 +1,21 @@
 package com.kbtg.bootcamp.posttest.exception;
 
+import lombok.Getter;
 import org.springframework.http.HttpStatus;
 
 import java.time.ZonedDateTime;
 
+@Getter
 public class ApiExceptionResponse {
-    private final String message;
-    private final HttpStatus httpStatus;
-    private final ZonedDateTime dateTime;
+	private final String message;
+	private final HttpStatus httpStatus;
+	private final ZonedDateTime dateTime;
 
-    public ApiExceptionResponse(String message, HttpStatus httpStatus, ZonedDateTime dateTime) {
-        this.message = message;
-        this.httpStatus = httpStatus;
-        this.dateTime = dateTime;
-    }
+	public ApiExceptionResponse(String message, HttpStatus httpStatus, ZonedDateTime dateTime) {
+		this.message = message;
+		this.httpStatus = httpStatus;
+		this.dateTime = dateTime;
+	}
 
-    public String getMessage() {
-        return message;
-    }
-
-    public HttpStatus getHttpStatus() {
-        return httpStatus;
-    }
-
-    public ZonedDateTime getDateTime() {
-        return dateTime;
-    }
 }
