+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/exception/BadRequestException.java
@@ -2,7 +2,7 @@ package com.kbtg.bootcamp.posttest.exception;
 
 public class BadRequestException extends RuntimeException{
 
-    public BadRequestException(String message) {
-        super(message);
-    }
+	public BadRequestException(String message) {
+		super(message);
+	}
 }
