+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/exception/InternalServiceException.java
@@ -2,7 +2,7 @@ package com.kbtg.bootcamp.posttest.exception;
 
 public class InternalServiceException extends RuntimeException{
 
-        public InternalServiceException(String message) {
-            super(message);
-        }
+		public InternalServiceException(String message) {
+			super(message);
+		}
 }
