+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/exception/NotFoundException.java
@@ -2,7 +2,7 @@ package com.kbtg.bootcamp.posttest.exception;
 
 public class NotFoundException extends RuntimeException{
 
-        public NotFoundException(String message) {
-            super(message);
-        }
+		public NotFoundException(String message) {
+			super(message);
+		}
 }
