+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/user/request/UserBuySellRequestDto.java
@@ -9,18 +9,11 @@ import lombok.Setter;
 @Getter
 public class UserBuySellRequestDto {
 
-    @NotNull
-    @Size (min = 10, max = 10, message = "userId must be 10 digits")
-    private  String userId;
+	@NotNull
+	@Size (min = 10, max = 10, message = "userId must be 10 digits")
+	private  String userId;
 
-    @NotNull
-    @Size (min = 6, max = 6, message = "ticketId must be 6 digits")
-    private  String ticketId;
-    public UserBuySellRequestDto() {
-    }
-
-    public UserBuySellRequestDto(String userId, String ticketId) {
-        this.userId = userId;
-        this.ticketId = ticketId;
-    }
+	@NotNull
+	@Size (min = 6, max = 6, message = "ticketId must be 6 digits")
+	private  String ticketId;
 }
