+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/user/service/UserService.java
@@ -19,92 +19,92 @@ import java.util.List;
 @Service
 public class UserService {
 
-    private final LotteryRepository lotteryRepository;
-    private final UserTicketRepository userTicketRepository;
+	private final LotteryRepository lotteryRepository;
+	private final UserTicketRepository userTicketRepository;
 
-    public UserService(LotteryRepository lotteryRepository, UserTicketRepository userTicketRepository) {
-        this.lotteryRepository = lotteryRepository;
-        this.userTicketRepository = userTicketRepository;
-    }
+	public UserService(LotteryRepository lotteryRepository, UserTicketRepository userTicketRepository) {
+		this.lotteryRepository = lotteryRepository;
+		this.userTicketRepository = userTicketRepository;
+	}
 
-    // Get all lotteries
-    @Transactional
-    public UserGetAllLotteriesResponse getAllLotteries() {
+	// Get all lotteries
+	@Transactional
+	public UserGetAllLotteriesResponse getAllLotteries() {
 
-        List<Lottery> lotteries = lotteryRepository.findAll();
-        if (lotteries.isEmpty()) {
-            throw new NotFoundException("No lottery available in the system");
-        }
+		List<Lottery> lotteries = lotteryRepository.findAll();
+		if (lotteries.isEmpty()) {
+			throw new NotFoundException("No lottery available in the system");
+		}
 
-        List<String> ticketIds = new ArrayList<>();
-        lotteries.forEach(lottery -> ticketIds.add(lottery.getTicketid()));
+		List<String> ticketIds = new ArrayList<>();
+		lotteries.forEach(lottery -> ticketIds.add(lottery.getTicketid()));
 
-        UserGetAllLotteriesResponse response = new UserGetAllLotteriesResponse();
-        response.setTickets(ticketIds);
+		UserGetAllLotteriesResponse response = new UserGetAllLotteriesResponse();
+		response.setTickets(ticketIds);
 
-        return response;
+		return response;
 
-    }
+	}
 
-    // Buy lotteries
-    @Transactional
-    public UserBuyLotteriesResponse buyLotteries(UserBuySellRequestDto requestDto) {
+	// Buy lotteries
+	@Transactional
+	public UserBuyLotteriesResponse buyLotteries(UserBuySellRequestDto requestDto) {
 
-        Lottery lottery = lotteryRepository.findByTicketid(requestDto.getTicketId());
-        if (lottery == null) {
-            throw new NotFoundException("Ticket not found");
-        }
+		Lottery lottery = lotteryRepository.findByTicketid(requestDto.getTicketId());
+		if (lottery == null) {
+			throw new NotFoundException("Ticket not found");
+		}
 
-        UserTicket transaction = new UserTicket(requestDto.getUserId(),
-                requestDto.getTicketId(), lottery.getPrice());
-        userTicketRepository.save(transaction);
+		UserTicket transaction = new UserTicket(requestDto.getUserId(),
+				requestDto.getTicketId(), lottery.getPrice());
+		userTicketRepository.save(transaction);
 
-        lottery.setAmount(lottery.getAmount() - 1);
+		lottery.setAmount(lottery.getAmount() - 1);
 
-        if (lottery.getAmount() <= 0) {
-            lotteryRepository.delete(lottery);
-        }
+		if (lottery.getAmount() <= 0) {
+			lotteryRepository.delete(lottery);
+		}
 
-        return new UserBuyLotteriesResponse(transaction.getId());
-    }
+		return new UserBuyLotteriesResponse(transaction.getId());
+	}
 
-    // Get user lotteries
-    @Transactional
-    public UserGetMyLotteriesResponse getMyLotteries(String userId) {
+	// Get user lotteries
+	@Transactional
+	public UserGetMyLotteriesResponse getMyLotteries(String userId) {
 
-        List<UserTicket> userTickets = userTicketRepository.findByUserId(userId);
+		List<UserTicket> userTickets = userTicketRepository.findByUserId(userId);
 
-        if (userTickets.isEmpty()) {
-            throw  new NotFoundException("User does not have any ticket or not found user id");
-        }
+		if (userTickets.isEmpty()) {
+			throw  new NotFoundException("User does not have any ticket or not found user id");
+		}
 
-        List<String> currentuserTickets = new ArrayList<>();
-        userTickets.forEach(userTicket -> currentuserTickets.add(userTicket.getTicketid()));
+		List<String> currentuserTickets = new ArrayList<>();
+		userTickets.forEach(userTicket -> currentuserTickets.add(userTicket.getTicketid()));
 
-        Integer totalAmount = 0;
-        for (UserTicket userTicket : userTickets) {
-            totalAmount += userTicket.getPrice();
-        }
+		Integer totalAmount = 0;
+		for (UserTicket userTicket : userTickets) {
+			totalAmount += userTicket.getPrice();
+		}
 
-        return new UserGetMyLotteriesResponse(currentuserTickets, totalAmount, userTickets.size());
-    }
+		return new UserGetMyLotteriesResponse(currentuserTickets, totalAmount, userTickets.size());
+	}
 
-    // Sell back my lottery ticket
-    @Transactional
-    public UserSellLotteriesResponse sellLotteries(UserBuySellRequestDto requestDto) {
+	// Sell back my lottery ticket
+	@Transactional
+	public UserSellLotteriesResponse sellLotteries(UserBuySellRequestDto requestDto) {
 
-        String userId = requestDto.getUserId();
-        String ticketId = requestDto.getTicketId();
+		String userId = requestDto.getUserId();
+		String ticketId = requestDto.getTicketId();
 
-        // find ticket id
-        UserTicket userTicket = userTicketRepository.findByUserIdAndTicketId(userId, ticketId);
-        if (userTicket == null) {
-            throw new NotFoundException("Ticket not found in user ticket list");
-        }
+		// find ticket id
+		UserTicket userTicket = userTicketRepository.findByUserIdAndTicketId(userId, ticketId);
+		if (userTicket == null) {
+			throw new NotFoundException("Ticket not found in user ticket list");
+		}
 
-        userTicketRepository.delete(userTicket);
+		userTicketRepository.delete(userTicket);
 
-        return new UserSellLotteriesResponse(ticketId);
-    }
+		return new UserSellLotteriesResponse(ticketId);
+	}
 
 }
