--- posttest/src/test/java/com/kbtg/bootcamp/posttest/PosttestApplicationTests.java
@@ -1,13 +1,0 @@
-package com.kbtg.bootcamp.posttest;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PosttestApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
